change: update pca_from_streamlit template to include conversion of results
The results file that is expected by a lot of functions is the one output
from the ProcessedResults class and therefore stores data as different
parts of the class. When you load data from the streamlit as part of an
H5 file, it gets loaded as a dict, so I convert it here so that you have both available to you.