add: compare to sae features change: ensure on mps as much as possible add: test of whether this works with day of the week features
I have tested this in GPT-2 with day of the week features. If you train it on Monday, the top feature by cosine similarity with the neurons in the linear probe is Monday, and the other most similar features are the other days of the week. I think this is quite a nice proof of concept.